err in stop

+++ itest/wasmd_handler.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"runtime"
 	"strconv"
 	"testing"
 	"time"
@@ -61,7 +62,7 @@ func (w *WasmdNodeHandler) Start() error {
 }
 
 func (w *WasmdNodeHandler) Stop(t *testing.T) {
-	err := w.stop()
+	err := w.stop(t)
 	if err != nil {
 		log.Printf("error stopping wasmd: %v", err)
 	}
@@ -104,7 +105,7 @@ func (w *WasmdNodeHandler) start() error {
 	return nil
 }
 
-func (w *WasmdNodeHandler) stop() (err error) {
+func (w *WasmdNodeHandler) stop(t *testing.T) (err error) {
 	if w.cmd == nil || w.cmd.Process == nil {
 		// return if not properly initialized
 		// or error starting the process
@@ -113,14 +114,15 @@ func (w *WasmdNodeHandler) stop() (err error) {
 
 	defer func() {
 		err = w.cmd.Wait()
+		if err != nil {
+			t.Logf("error waiting for wasmd process to exit: %v", err)
+		}
 	}()
 
-	return w.cmd.Process.Signal(os.Kill)
-
-	//if runtime.GOOS == "windows" {
-	//	return w.cmd.Process.Signal(os.Kill)
-	//}
-	//return w.cmd.Process.Signal(os.Interrupt)
+	if runtime.GOOS == "windows" {
+		return w.cmd.Process.Signal(os.Kill)
+	}
+	return w.cmd.Process.Signal(os.Interrupt)
 }
 
 func (w *WasmdNodeHandler) cleanup() error {
